Extract Unix file status partial
